requirements.txt
aiogram
aiosqlite
discord.py